Add deprecation notices
Histogram2D is replaced by just using Chart2D and templating
is conditional on an IHistogramRenderData instance being present.
Customizable 2D projection is going to be removed due to causing
incompatibility with the host application. Long term it will be better
to always standardize to the default XZ projection, and instead make
more wrapper utility methods to preform automatic creation and
conversion if other forms are desired.